Remove dead code, move scalar color oracles into tests, switch prints to logging
- Delete unused mesh.compute_edge_map, video.pixelate_frame, colors.main()
- Move per-cell reference color functions to tests/reference_colors.py; they
  only served as correctness oracles for the vectorized downsample paths
- Replace stdout prints in video.py/mesh.py with module loggers
